The Maximum String
here we defined two variable currentSum=0 and maxSum =nums[0], then we ran a loop and with the reach cycle we added the ith position of the vector in the currentSum and if the current sum is bigger than the maxSum then we will store it in the maxSum , and if the currentSum is <0 we will convert the currentSum to 0 (because now we know that it will not contribute in the maxSUm and we also do not have to worry that what if the max was in the array which is already processed in currentSum but due to some -ve no. it become -ve like [1,2,3,4,5,-50,5,4] the [1,2,3,4,5] part of the array is already been given to maxSum and now no other value of currentSum will be able to replace this from the maxSum ...because its the maximum.