287. Find the Duplicate Number
here i have done it in two ways 1st i use a very long array (vector) and then see where the index of the vector in explored or not , if yes then return the value as the answer if not then mark the index as explored.

the 2nd way is by using fast pointer and slow pointer linked list cycle detection method .. in this ,method i have to pointer , on jump 2 points and one jump only 1 point. if there is a cycle then they are destined to meet . and if they meet send one pointer to the start and convert the both pointer into slow pointer meaning they will only skip only on point, and the next time when they will meet ,it will ne point t which we want.