Find the repeating and missing numbers
here 1st we will create a variable and there we will store the xor of the array and then we for the 1-n so now we will have the xor of one dupricate no. and missing no. ,then we will find the right most bit by "z = xor1 & ~(xor1 - 1);" and then we will make two variable x and y and if array[i]^z==0 we will p x^= aray[i] else p y^= aray[i]....and this x and y will give the our requited no.